// apps/backend/src/generated/models/BasicPersonDetailsDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BasicPersonDetailsDTO = {
    /**
     * Family name of the person
     */
    familyName?: string | null;
    /**
     * First name known as
     */
    firstNameKnownAs?: string | null;
    /**
     * Given name of the person
     */
    givenName?: string | null;
    /**
     * Initials of the person
     */
    initials?: string | null;
    /**
     * Title of the person
     */
    title?: string | null;
    /**
     * User number of the person
     */
    userNumber?: string;
    /**
     * Organisation name of the person of which the person is part of
     */
    orgName?: string | null;
    /**
     * Organisation acronym
     */
    orgAcronym?: string | null;
    /**
     * Department name of which the person is part of
     */
    deptName?: string | null;
    /**
     * Department acronym
     */
    deptAcronym?: string | null;
    /**
     * Country of the person
     */
    country?: string | null;
    /**
     * Email of the person
     */
    email?: string | null;
    /**
     * Work phone of the person
     */
    workPhone?: string | null;
    /**
     * Establishment ID of the person
     */
    establishmentId?: string | null;
    /**
     * Full name of the person
     */
    fullName?: string | null;
    /**
     * Display name of the person
     */
    displayName?: string | null;
};